Fix Guid serialization, fix failing Guid polyglot test
